+++ src/ui/components/download-button.tsx
@@ -3,27 +3,26 @@ import { createLink } from '../utils/links.ts'
 import { IconButton } from './icon-button.tsx'
 import type { PropsWithChildren, ReactElement } from 'react'
 
-export interface DownloadBlockButtonProps extends PropsWithChildren {
+export interface DownloadButtonProps extends PropsWithChildren {
   ipfsPath: string
   className?: string
 }
 
-export function DownloadBlockButton ({ ipfsPath, children, className }: DownloadBlockButtonProps): ReactElement {
-  function downloadBlock (evt: MouseEvent): void {
+export function DownloadButton ({ ipfsPath, children, className }: DownloadButtonProps): ReactElement {
+  function download (evt: MouseEvent): void {
     evt.preventDefault()
     evt.stopPropagation()
 
     window.location.href = createLink({
       ipfsPath,
       params: {
-        format: 'raw',
         download: 'true'
       }
     })
   }
 
   return (
-    <IconButton className={className} onClick={downloadBlock} title='Download block'>
+    <IconButton className={className} onClick={download} title='Download'>
       {children}
     </IconButton>
   )