+++ test-e2e/dnslink.test.ts
@@ -60,7 +60,7 @@ test.describe('DNSLink', () => {
 
     const headers = await response.allHeaders()
     expect(headers?.['content-type']).toContain('application/vnd.ipld.dag-json')
-    expect(headers?.['cache-control']).toBe('public, max-age=60')
+    expect(headers?.['cache-control']).toMatch(/public, max-age=60, stale-while-revalidate=\d+, stale-if-error=\d+/)
     expect(headers?.['etag']).toBe(`"${cid}.dag-json"`)
     expect(headers?.['x-ipfs-path']).toBe(`/ipns/${domain}`)